Refactor FlyingEnemy script for null check optimization
Updated the FlyingEnemy.cs script to optimize the null checks on m_WaypointMover and _spriteRenderer. The null checks are now done once at the Start method instead of checking every time during runtime in the CheckFlipSprite and SetStateInformation method. This change will help improve the performance by reducing the unnecessary checks. Changed waypointMover variable to m_WaypointMover to adhere to the coding standards.