Add static classes and refactor namespaces across scripts
This commit introduces static instance classes(StaticInstance, Singleton, and PersistentSingleton) in the Utility directory for better reusable code. Several scripts are also updated to reflect the new directory structure to be consistent. Furthermore, the QualitySettings.asset is updated to align with the latest Unity versions. The playerInputs.inputactions method of code generation has been enabled for better utilization of Unity's input system.