Refactor UIManager and WalkingEnemy scripts
In UIManager.cs, InputManager is now obtained directly from an instance, removing the need for the SetUpInputManager() method. This improves the overall architecture and makes code more maintainable. The method CheckPauseInput() is simplified and transformed into an event response method, namely OnPause(), providing a more intuitive understanding of its functionality. As a result, some Tooltips and functionalities irrelevant after the changes were removed.

In WalkingEnemy.cs, we updated the namespaces to ensure clearer and more accurate categorization.