Update Game UI Canvas prefab and fix enemy behaviour
This commit consists of several updates, mainly focused on the Game UI Canvas prefab, adjusting the scaling and navigation settings. The changes make sure that the proportions scale consistently in all child elements, plus a navigation wrapping was disabled to ensure the UI shows correctly on screens.

Additionally, this commit removes a deprecated field `enemyState` in the `BasicWalkingEnemy.prefab`. With reference to the `BasicWalkingEnemyAnimatorController`, the current animation state is better managed, hence the field became unnecessary.

Previously, the paths of the scenes in the `EditorBuildSettings.asset` were pointing to the wrong location. This was corrected by redirecting the paths to the appropriate ones under the `Prefab` directory.

Lastly, the string argument in `MainMenu` was changed to `Level2` to improve user navigation, and the default selected component in `BasicWalkingEnemy.prefab` was updated for better interaction management.