Remove obsolete InputSystem settings and level
Deleted file 'InputSystem.inputsettings.asset', its metadata, and 'Level1.unity'. These files were obsolete and were no longer relevant in the context of the current setting. Majority of the data in these files were either not used or depricated and therefore, was resulting in redundancy and unnecessary memory use. Keeping these would be a potential source of confusion for future maintenance.