Remove unused assets, update player scripts
This commit cleans up the project by removing unused assets and updating the player scripts for better functionality. The PlayerController script was heavily modified to improve character movement, additionally new scripts for PlayerCharacter and corrections in PlayerAnimator were introduced. A new MonoBehaviour GameManager is introduced, it removes the usage of Singleton from the PlayerController reducing the project's complexity. The collisions with BoxCollider2D were improved by adding checks for grounded state and for triggers. In UIManager, the InputManager reference was removed, since it was unused. The reference guid for PlayerController.cs and PlayerAnimator.cs were modified due to script changes. These changes improve the overall functionality and performance of the project while simplifying the structure.