Audio Manager Prefab is created and AudioManager.cs is updated
Old version of the AudioManager was taking a track number as an input. After some thinking I thought what if I just pass an enum instead of dealing with all the track orders in Arrays. So here we are with an enum based AudioManager that is scalable with the ease of use.