InputManager Class extended and its architecture is changed
Previous version works with setting public fields depending on the input, when a class needed an input it was just accessing the public field and using it. Since this system has so much coupling and it was prone to error after a certain scale. I changed it with a event based system. Now Every class can just listen to inputs without the need of  public fields.