Ground Check Base class, EnemyGroundCheck, and WallColliderCheck are created
Old ground check was working with rays which wasn't really accurate and efficient. Now its based on  OnTriggerExit and OnTriggerEnter which is much more reliable and efficient(I believe so) since they are only able to collide with ground layer which decreases the amount of collisions to check each frame.