PlayerStayingOnHead.cs is created
This solves a somewhat interesting problem. Normally Enemy would always turn towards the player but this creates the issue of non stop turning since player is always in right or left with a floating point difference. So instead of making the enemy dizzy with turning it nonstop(such a cruel thing to do) I created an additional check for this.